I added two new requirements printers: - IdRequirementsPrinter - InventoryValueRequirementsPrinter
Furthermore, I added two new validator classes:
- IdValidator
- InventoryValueValidator

Finally, I changed the Product file by transforming the return value from double to long double.

+++ src/Product.cpp
@@ -50,7 +50,7 @@ void Product::setStock(int stock) {
 	return stock_;
 }
 
-[[nodiscard]] double Product::calculateInventoryValue() const noexcept {
+[[nodiscard]] long double Product::calculateInventoryValue() const noexcept {
 	return unitPrice_ * stock_;
 }
